requirements.txt
Flask==3.1.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
playwright==1.51.0
